+++ CHANGELOG.md
@@ -8,6 +8,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.4.0-pre.1] - 2025-10-10
+
+### Added
+
+- Added support for CPU camera image capture, and camera display-related render matrices.
+
+### Changed
+
+- Changed AR Foundation dependency version from 6.3.0 to 6.4.0-pre.1.
+
 ## [2.3.0] - 2025-10-03
 
 ### Changed