+++ Documentation~/whats-new.md
@@ -1,10 +1,10 @@
 ---
 uid: meta-openxr-whats-new
 ---
-# What's new in version 2.3
+# What's new in version 2.4
 
-## Environment ray casts
+## CPU Camera Images
 
-* Added support for ray casts against the **Depth** trackable type via `XR_META_environment_raycast`. Refer to [Ray casts](xref:meta-openxr-raycasts) for more information.
+* Added support for CPU camera image capture, and camera display-related render matrices.  Refer to [Camera Image Capture](xref:meta-openxr-camera) for more information.
 
 For a full list of changes in this version including backwards-compatible bugfixes, refer to the package [changelog](xref:meta-openxr-changelog).